pyproject.toml
[tool.poetry]
name = "oshin-lab-infra"
version = "0.1.0"
description = ""
authors = ["shingoasou-0804 <122670383+shingoasou-0804@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.41.1"
pyrebase4 = "^4.8.0"
python-dotenv = "^1.0.1"
setuptools = "^75.8.0"
google-cloud-bigquery = {extras = ["bqstorage", "pandas"], version = "^3.30.0"}
pandas = "^2.2.3"
db-dtypes = "^1.4.1"
langchain = "^0.3.20"
langchain-openai = "^0.3.8"
traceloop-sdk = "^0.38.12"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.10.0"
flake8 = "^7.1.1"
pip-audit = "^2.7.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
